Move migrator elsewhere, use dialect directory
